+++ CHANGES.md
@@ -2,6 +2,7 @@
 * Fixed export creating JSON with null values for Styles that have parents
 * Allow drag and drop of SCMP files into main screen to load project.
 * Created export dialog and moved export settings to this window.
+* Improved appearance of Settings dialog.
 
 ### Skin Composer Version 25 ###
 * Fixed exception when opening an old project.